build in custom image (#46)
Switching to building in this image: https://github.com/signalflatpak/image instead of debian:bookworm. It's a debian 13 image that will include the dependencies we're installing anyway.

This should save a little bit of build time. Docker was also down yesterday when 7.72 was building so it couldn't pull the bookworm image from docker hub, and this should solve that potential problem too.